ml-trading-engine/requirements.txt
pandas
numpy
xgboost
scikit-learn
ta
joblib
pyyaml
schedule
matplotlib